requirements.txt
faster-whisper==1.0.3
sounddevice==0.4.6
numpy==1.26.4
customtkinter==5.2.2
pillow==10.2.0
av==12.0.0
requests==2.31.0